refactor: 페이지네이션이 올바르게 작동하도록

+++ src/main/java/com/inha/capstone/recommendation/dto/RecommendationEventListResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record RecommendationEventListResponse(
     List<RecommendationEventResponse> items,
-    int total,
+    long total,
     int page,
     int pageSize
 ) {}
